# backend/src/main/java/com/platform/user/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: User Module RBAC API
  description: |
    Comprehensive Role-Based Access Control (RBAC) API for the Spring Boot Modulith payment platform.

    This API provides multi-tenant permission management with organization isolation, custom role creation,
    and real-time permission checking with sub-200ms performance requirements.

    ## Features
    - Multi-tenant organization isolation
    - System and custom role management
    - Resource×Action permission model
    - Role assignment with expiration support
    - Comprehensive caching for performance
    - Event-driven architecture

    ## Authentication
    All endpoints require valid JWT authentication with appropriate permissions.

    ## Rate Limiting
    - Permission checks: 1000 requests/minute per user
    - Role management: 100 requests/minute per user
    - User assignments: 500 requests/minute per user
  version: 1.0.0
  contact:
    name: Platform Team
    email: platform@company.com
  license:
    name: Proprietary
    url: https://company.com/license
servers:
  - url: https://api.platform.com/v1
    description: Production server
  - url: https://staging-api.platform.com/v1
    description: Staging server
  - url: http://localhost:8080
    description: Local development server
security:
  - BearerAuth: []
...
paths:
  /api/permissions:
    get:
      tags: [Permissions]
      summary: Get all system permissions
      description: |
        Retrieve all available system permissions. Used for permission management
        interfaces and role creation.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all system permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/permissions/resources:
    get:
      tags: [Permissions]
      summary: Get all permission resources
      description: Get all unique resources in the permission system
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of permission resources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcesResponse'
  /api/permissions/resources/{resource}/actions:
    get:
      tags: [Permissions]
      summary: Get actions for a resource
      description: Get all actions available for a specific resource
      parameters:
        - name: resource
          in: path
          required: true
          schema:
            type: string
            example: PAYMENTS
      responses:
        '200':
          description: List of actions for the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionsResponse'
  /api/organizations/{organizationId}/permissions/check:
    post:
      tags: [Permissions]
      summary: Check multiple permissions
      description: |
        Check multiple permissions for the current user in an organization.
        Efficient batch permission checking for complex authorization scenarios.
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionCheckRequest'
      responses:
        '200':
          description: Permission check results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionCheckResponse'
    get:
      tags: [Permissions]
      summary: Check single permission
      description: Check a single permission for the current user
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - name: resource
          in: query
          required: true
          schema:
            type: string
            example: PAYMENTS
        - name: action
          in: query
          required: true
          schema:
            type: string
            example: WRITE
      responses:
        '200':
          description: Single permission check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SinglePermissionCheckResponse'
  /api/organizations/{organizationId}/permissions/my-permissions:
    get:
      tags: [Permissions]
      summary: Get current user permissions
      description: Get all permissions assigned to the current user in an organization
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      responses:
        '200':
          description: User permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPermissionsResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        example: 1
      description: Organization identifier
...
  schemas:
    PermissionInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        resource:
          type: string
          example: PAYMENTS
        action:
          type: string
          example: WRITE
        description:
          type: string
          example: "Write access to payment operations"
      required:
        - id
        - resource
        - action
    PermissionsResponse:
      type: object
      properties:
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionInfo'
    ResourcesResponse:
      type: object
      properties:
        resources:
          type: array
          items:
            type: string
          example: ["USERS", "PAYMENTS", "ORGANIZATIONS", "SUBSCRIPTIONS", "AUDIT"]
    ActionsResponse:
      type: object
      properties:
        resource:
          type: string
          example: PAYMENTS
        actions:
          type: array
          items:
            type: string
          example: ["READ", "WRITE", "ADMIN"]
    PermissionCheckRequestItem:
      type: object
      properties:
        resource:
          type: string
          example: PAYMENTS
        action:
          type: string
          example: READ
      required:
        - resource
        - action
    PermissionCheckRequest:
      type: object
      properties:
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionCheckRequestItem'
      required:
        - permissions
    PermissionCheckResult:
      type: object
      properties:
        resource:
          type: string
          example: PAYMENTS
        action:
          type: string
          example: READ
        hasPermission:
          type: boolean
          example: true
    PermissionCheckResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/PermissionCheckResult'
    SinglePermissionCheckResponse:
      type: object
      properties:
        resource:
          type: string
          example: PAYMENTS
        action:
          type: string
          example: WRITE
        hasPermission:
          type: boolean
          example: false
    UserPermissionsResponse:
      type: object
      properties:
        permissions:
          type: array
          items:
            type: string
          example: ["PAYMENTS:READ", "PAYMENTS:WRITE", "USERS:READ"]
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Invalid request parameters"
        correlationId:
          type: string
          example: "abc123-def456-ghi789"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...